fix clippy::unnecessary_cast

+++ rustysynth/src/modulation_envelope.rs
@@ -90,9 +90,9 @@ impl ModulationEnvelope {
 
         while self.stage <= EnvelopeStage::HOLD {
             let end_time = match self.stage {
-                EnvelopeStage::DELAY => self.attack_start_time as f64,
-                EnvelopeStage::ATTACK => self.hold_start_time as f64,
-                EnvelopeStage::HOLD => self.decay_start_time as f64,
+                EnvelopeStage::DELAY => self.attack_start_time,
+                EnvelopeStage::ATTACK => self.hold_start_time,
+                EnvelopeStage::HOLD => self.decay_start_time,
                 _ => panic!("Invalid envelope stage."),
             };
